Adding EntityStatMod System
Added a system to handle adding modifiers to any entity stat. These exist as resources and can be added, updated, or deleted. They can also stacked if specified. Work is also beginning on stats that change how damage is received based on what kind of damage it is (with the EffectModifierComponent).