Refactoring Stat Mod & Status Effect Systems
Attempted to simplify the source of truth for both stat mods and current effects on entities by moving the caches and dictionaries of active effects to the entity itself... rather than it being spread out amongst the nodes. There is some error-prevention in place with assert statements as of now, but it would be a good idea to keep flushing out its simplicity in the future.