Adding Auto Stacking and Dragging Half Stacks
Added the ability to auto stack stray items in the inventory that aren't at stack size yet. It compacts them into the lowest inventory array index. Also added the ability to drag half a stack by holding the sprint key (shift for now) and right clicking. This effectively splits a stack. Also added the ability to double click on an item to attempt to garner it into a full stack from items in the surrounding slots.