Implementing Item YSorts & Auto Pickup Delay
Added y sorting for the testing nodes like items so that they draw in correct layered order related to their y position. Also added a delay for when items get spawned and when they can be auto picked up. This fixes the issue where an auto-pickup item got dropped and instantly readded.