Adding Storm Transform Resources & Fixing Noise
Added a new resource that defines how and where the storm should move to. The storm script now has a storm transform queue that moves it in phases based on resources in the queue. A bug with the storm noise moving with the zone has been fixed, so now it simulates staying in place.