Adding Oak & Spruce Logs and Two Handed Icons
Added spruce and oak log art and resources. Added a new item type called "WorldResource". Also added the ability to have a non-animated off hand sprite created (with a custom offset & rotation) should the item need it. This works well with the big logs that would be held with two static hands.