Adding Entity Footsteps, Streaks, & Popup Numbers
Added footsteps that get placed then fade out where the entity has been. Also added a streak in the ground at the feet of the entity for when they are knocked back. Added popup numbers for damage, healing, and crits. Fixed a bug where reload ammo was being consumed twice.