Reworking HotbarUI & Adding Steel Boomerang
This is the beginning of a rework for the hotbar UI and general slot UI. The active slot now grows and changes texture. Item slots now have a color behind them indicating rarity instead of the outline. Also added the steel boomerang and gave it the ability to chop trees.