Finishing Crafting UI & Fixing Rarity Glow
Finished the crafting system UI and rarity checks. Added a button to pull the craft out and reworked logic for dropping extras on the ground. Also fixed a bug with the item interaction hud after quantities were combined. Fixed the rarity glow flicker by toning it down and only glowing the non-outlined pixels.