Adding Flintlock & Rocks
Added a flintlock weapon with a reload animation similar to the blunderbuss. Also added rocks to the scene for at least one more thing to shoot at. Finally, the weapon clipping detection was reworked to happen on a separate layer with a specific area2d on each object that can trigger the clip.