Starting to Implement Wearables
Laying the ground work for implementing wearables. These will provide stat buffs or nerfs to the entity wearing it. The next step will be to create the UI for it, similar to how the weapon mod UI works. Also, the scaling for resistances and weaknesses for things like burning was changed to a 0 to 100 scale, with 0 as the default.