Refactoring Interaction HUD & Fixing Bugs
Refactored the interaction HUD to make it consistent and easy to trigger. Removed the specific item pickup hud in favor for the new system. Fixed some bugs regarding hotbar sizes in the inventory and made it a constant. There is a known bug that items that get crafted are not unique.