Adding AOE FX, Trails, & Fixing Proj Arcing Bugs
Added AOE ground FX and flames out of the bottle in the air. Also fixed a bug where arcing and splitting fire bottles would keep moving after hit. Also fixed a bug with the confusion status effect where it was trying to require a handler for particles but none should exist in the first place. Finally, trails to projectiles were added, notably to boomerangs.