Refactoring Audio Funcs & Recusive Caching
Refactored the functions to start audio. Now they are simpler with 2d and global being the two func options. Also added recursive caching to the TRES folder so that it can have subfolders of items for better organization (and still be cached). There is also a way to keep sounds in memory by registering and unregistering them. It works great for rapid use of firing sounds.