Refactoring FSM States & Improving Audio Workflow
Refactored FSM states to be resource based, which is lighter on memory and cpu than nodes, but is a pain to ensure uniqueness. Also extracted most of the running/sprinting logic to a set of static methods that can operate on any run/sprint based state. Finally, the workflow involving holding references to an audio player has been greatly improved with a new validity checker method.