Improving Enemy AI & Fixing Die/Spawn States
Improved local obstacle avoidance for enemies. Simplified entity inheritance with a new Entity class, which will be extra useful in cleaning up scripts after a save system rework. Also fixed the die and spawn states, making things stop collide during death and actually freeing it if it is not a player.