Reworking Melee Sequence & Fixing Bugs
Reworked the melee weapon script to include several of the improvements made to the projectile weapon script a few commits ago. Also fixed bugs with the overhead ui and cursor tint. Added new charge bar to show how far along a charge is, and made charges deplete instead of go instantly to 0 on release.