Adding Btn Icons & Reworking Audio Instances
Added icons to all UI buttons to replace the text. Also Reworked the way audio players are instanced by extending them with a custom type. This solves a bug with players reentering the tree but with a new sound and still being tracked by a node that held a reference.