Starting Proj Refactor, Improving Wpn Lvl Up Logic
Started the projectile refactoring by splitting up some of the file, but the logic is still the same for now. Also improved the level up logic for weapons, consolidating the methods into one. Still need a way to notify and handle level ups, potentially for increasing damage, etc. Also notably, the font spacing issue was fixed by reimporting the font file with auto mipmapping on.