Adding Context Tooltips & Boompowder
Added tooltips to many drag and drop actions. Upgraded the colors in Globals and added glowing options for better game feel. Added boompowder as a new ammo type, replacing what was to be "rockets". Redid the art for ion charges. Added new sets of slots that are a WIP for ammo and materials of some kind (maybe gold, or building mats, etc.)